% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pdf.R
\name{parse_pdf}
\alias{parse_pdf}
\title{An R wrapper for PDF form extraction}
\usage{
parse_pdf(filenames, outfile = NULL)
}
\arguments{
\item{filenames}{Path(s) to the PDF files to extract from}

\item{outfile}{Path to csv file output (optional)}
}
\value{
tibble of dim [n,m] for n files and m fields
}
\description{
Parse the Field data from PDF forms and store output into a tibble.
Optionally write out to csv file as well.
}
