% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pdftotext_layout_output.R
\name{read_bbox_layout_xhtml}
\alias{read_bbox_layout_xhtml}
\title{Read PDF element positions from {pdftotext}.}
\usage{
read_bbox_layout_xhtml(path_to_html)
}
\arguments{
\item{path_to_html}{Path to HTML file generated by {pdftotext}.}
}
\description{
\code{read_bbox_layout_xhtml} Parses PDF layout files produced
by \code{pdftotext} command \code{pdftotext file -bbox-layout}.
The function returns a tibble with bounding box information for
each word, line, block and page.
}
\examples{
doc <- system.file("extdata", "edi_2009_frcho43c6mmlx5lyohqy_doc#immrrkosg.html", package = "pdfparser")
read_bbox_layout_xhtml(doc)
}
