% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pdftotext_layout_output.R
\name{read_pdf_layout}
\alias{read_pdf_layout}
\title{Read PDF element positions from {pdftotext}.}
\usage{
read_pdf_layout(path_to_html)
}
\arguments{
\item{path_to_html}{Path to HTML file generated by {pdftotext}.}
}
\description{
\code{read_pdf_layout} Parses PDF layout files produced
by the command line tool \code{pdftotext}.
The function returns a tibble with layout information for
each word, line, block and page.
}
\examples{
doc <- system.file("extdata", "edi_2009_frcho43c6mmlx5lyohqy_doc#immrrkosg.html", package = "pdfparser")
read_pdf_layout(doc)
}
